Refactor autogenerated client out of quetzal.client
Massive refactoring to put the autogenerated code outside the quetzal.client
module, which breaks the weird module organization that did not permit to create
meaningful synonyms/shortcuts on the quetzal/client/__init__.py file. One
pending thought to keep in mind: should the autogenerated client live in a
separate repo? For the moment, let's keep things together to avoid
fragmentation.

A fortunate side-effect of this refactoring is that the issue concerning the
dots on the generated
files (https://github.com/OpenAPITools/openapi-generator/issues/1658) has been
fixed; there is no need to manually rename files.

Finally, a namespace package distribution strategy was chosen and implemented
for the quetzal namespace. There are three strategies possible; the one used for
quetzal will be "pkgutil-style namespace packages". See:
https://packaging.python.org/guides/packaging-namespace-packages/#pkgutil-style-namespace-packages