Add dependency links
This commit is intended to fix the use-case of adding quetzal-client as a
dependency/requirement on a requirements.txt or through pip. It currently fails
but this seems to inform pip where to find the package.

When quetzal-* packages are added to PyPI, there will be no need for this.